+++ packages/ipydrawio/src/tokens.ts
@@ -46,6 +46,8 @@ export const DRAWIO_ICON_SVG = ICON_SVG;
 
 export const IPYDRAWIO_METADATA = NS;
 
+import SCHEMA from '../schema/plugin.json';
+
 /**
  * Escape hatch for runtime debugging.
  */
@@ -67,7 +69,7 @@ export const DRAWIO_ICON_CLASS_RE = /jp-icon-warn0/;
 export const DIAGRAM_MENU_RANK = 99;
 
 // TODO: this is duplicated in schema
-export const UI_THEMES = ['min', 'atlas', 'dark', 'kennedy'];
+export const UI_THEMES = SCHEMA.definitions['ui-theme']['enum'];
 
 export const IDiagramManager = new Token<IDiagramManager>(PLUGIN_ID);